fix: Path Limit Size Rule - Closer Match to API Standards Description (#43)
The description indicates the host + URL size can be "hundreds" of characters. As a result, increasing this value to 100 before the warning seems appropriate, as it is just a guide. This seems pretty long even at 100 but is more reasonably in line with the expected time to see this warning (when you do something extreme).