feat: Bulk Synchronous Standards (#69)
Proposal of new API Standards for bulk operations that attempt to
provide some consistency with a RESTful approach. That being said there
are many compromises here. This content covers concepts around
Synchronous bulk operations to a single resource.

Bulk operations to multiple resources is out of scope, along with async
bulk operations and async in general. That being said, schema for bulk
operations should be reviewed and considered for consistency in future
standards around async and schema.

---------

Signed-off-by: Travis Gosselin <travisgosselin@users.noreply.github.com>
Co-authored-by: Evan Gilbert <eggilbert@users.noreply.github.com>