feat(DPE-35): Created Collections rules (#86)
Impact testing results are about 8 more errors in total. The execution
time for the impact testing is within a second or so compared to the
current ruleset in the main branch.

Performance testing results are okay. These changes will create about
700 more errors or warnings within this Open API document suite we use.
About 564 of those 700 errors/warnings are mainly concentrated in a few
documents. The two rules `sps-missing-pagination-query-parameters` and
`sps-no-collection-paging-capability` are the ones that are causing the
most errors since not all GET endpoints are not designed with
collections capabilities. Execution time is still within a second or so
of the current ruleset on the main branch.

---------

Signed-off-by: Ethan Honzik <105084033+EthanHonzikSPS@users.noreply.github.com>
Co-authored-by: Ethan Honzik <ehonzik@spscommerce.com>
Co-authored-by: Ethan Honzik <105084033+EthanHonzikSPS@users.noreply.github.com>
Co-authored-by: Mark DeBeer <me@markdebeer.com>