Reconsider: Request / Response Execution Context Header Requirement (#108)
Clarified requirements for header values in requests and responses for
execution context. The original intent was that there may be scenarios
where the response value provides a different alias, or redirected
version that is available. This was worded carefully in the statement
before and after in the original content. Adding this line in the last
update overrides that intent.

---------

Signed-off-by: Travis Gosselin <travis.gosselin@live.ca>